requirements.txt
sentence-transformers>=2.2.0
numpy>=1.24.0,<2.0.0
scipy>=1.10.0
umap-learn>=0.5.0
matplotlib>=3.7.0
torch>=2.0.0
